requirements.txt
absl-py==2.1.0
accelerate==0.33.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
array_record==0.5.1
asttokens==3.0.0
astunparse==1.6.3
attrs==25.1.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
certifi==2024.7.4
-e git+ssh://git@github.com/guochengqian/openpoints.git@baeca5e319aa2e756d179e494469eb7f5ffd29f0#egg=chamfer&subdirectory=cpp/chamfer_dist
charset-normalizer==3.3.2
click==8.1.7
cmake==3.30.0
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.11
decorator==5.2.1
deepspeed==0.14.5
defusedxml==0.7.1
diffusers==0.11.1
dm-tree==0.1.8
docker-pycreds==0.4.0
docopt==0.6.2
docstring_parser==0.16
draccus==0.9.3
easydict==1.13
egl_probe==1.0.2
einops==0.8.0
einops-exts==0.0.4
emd-ext==0.0.0
etils==1.7.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
fpsample==0.3.2
fsspec==2024.6.1
ftfy==6.3.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
hjson==3.1.0
huggingface-hub==0.25.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
immutabledict==4.2.0
importlib_metadata==8.0.0
importlib_resources==6.4.0
ipython==8.12.3
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==3.4.1
kiwisolver==1.4.5
libclang==18.1.1
lit==18.1.8
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mergedeep==1.3.4
mistune==3.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multimethod==1.12
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.560.30
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
open_clip_torch==2.29.0
opencv-python==4.10.0.84
OpenEXR==3.2.4
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
peft==0.14.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.2.2
-e git+ssh://git@github.com/guochengqian/openpoints.git@baeca5e319aa2e756d179e494469eb7f5ffd29f0#egg=pointnet2_cuda&subdirectory=cpp/pointnet2_batch
pooch==1.8.2
prismatic==0.4
promise==2.3
prompt_toolkit==3.0.50
protobuf==3.20.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==16.1.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pyvista==0.44.1
PyYAML==6.0.1
pyyaml-include==1.4.1
pyzmq==26.2.1
referencing==0.36.2
regex==2024.5.15
requests==2.32.3
rich==13.7.1
-e git+ssh://git@github.com/droid-dataset/droid_policy_learning.git@10c5632733e7e297fd52ae7bca2232f6a2828993#egg=robomimic
rpds-py==0.23.1
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
scooby==0.10.0
sentry-sdk==2.9.0
setproctitle==1.3.3
shortuuid==1.0.13
simple_parsing==0.1.5
six==1.16.0
smmap==5.0.1
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.16.2
tensorflow-addons==0.23.0
tensorflow-datasets==4.9.6
tensorflow-graphics==2021.12.3
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.15.0
termcolor==2.4.0
threadpoolctl==3.5.0
timm==1.0.8
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
torch==2.0.1
torchvision==0.15.2
tornado==6.4.2
tqdm==4.66.4
traitlets==5.14.3
transformers==4.47.1
trimesh==4.4.2
triton==2.0.0
typeguard==2.13.3
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
vtk==9.3.1
wandb==0.17.4
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
wrapt==1.16.0
yarg==0.1.9
zipp==3.19.2